# lab5web/docs/swagger.yaml
definitions:
  handlers.ArticleCreateInput:
    properties:
      content:
        example: Текст статьи...
        type: string
      status:
        enum:
        - draft
        - published
        example: draft
        type: string
      title:
        example: Моя первая статья
        type: string
    required:
    - status
    - title
    type: object
  handlers.ArticleResponse:
    properties:
      content:
        example: Текст статьи...
        type: string
      created_at:
        example: "2024-01-15T12:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        example: published
        type: string
      title:
        example: Моя первая статья
        type: string
      updated_at:
        example: "2024-01-15T12:30:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  handlers.ArticleUpdateInput:
    properties:
      content:
        example: Обновленный текст...
        type: string
      status:
        enum:
        - draft
        - published
        example: published
        type: string
      title:
        example: Обновленный заголовок
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/articles:
    get:
      description: Возвращает все статьи текущего авторизованного пользователя с пагинацией
      parameters:
      - default: 1
        description: Номер страницы
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Количество записей на странице
        example: 10
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список статей
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Получение списка статей пользователя
      tags:
      - Articles
    post:
      consumes:
      - application/json
      description: Создает новую статью для авторизованного пользователя
      parameters:
      - description: Данные для создания статьи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ArticleCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Статья успешно создана
          schema:
            $ref: '#/definitions/handlers.ArticleResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Создание новой статьи
      tags:
      - Articles
  /api/articles/{id}:
    delete:
      description: Удаляет статью (soft delete) (только если пользователь является
        владельцем)
      parameters:
      - description: UUID статьи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Статья успешно удалена
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Доступ запрещен (не владелец)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Статья не найдена
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Удаление статьи
      tags:
      - Articles
    get:
      description: Возвращает статью по указанному ID (только если пользователь является
        владельцем)
      parameters:
      - description: UUID статьи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Статья найдена
          schema:
            $ref: '#/definitions/handlers.ArticleResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Доступ запрещен (не владелец)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Статья не найдена
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Получение статьи по ID
      tags:
      - Articles
    put:
      consumes:
      - application/json
      description: Обновляет существующую статью (только если пользователь является
        владельцем)
      parameters:
      - description: UUID статьи
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ArticleUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: Статья успешно обновлена
          schema:
            $ref: '#/definitions/handlers.ArticleResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Доступ запрещен (не владелец)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Статья не найдена
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Обновление статьи
      tags:
      - Articles
...
swagger: "2.0"